+++ src/main/php/web/handler/Frontend.class.php
@@ -17,7 +17,7 @@ class Frontend implements \web\Handler {
    *
    * @param   web.Request $request
    * @param   web.Response $response
-   * @return  void
+   * @return  var
    */
   public function handle($request, $response) {
     $action= $this->actions->from($request);