+++ src/test/php/web/unittest/filters/InvocationTest.class.php
@@ -72,4 +72,10 @@ class InvocationTest extends \unittest\TestCase {
     ])]);
     $fixture->proceed(new Request(new TestInput('GET', '/')), new Response(new TestOutput()));
   }
+
+  #[@test]
+  public function returns_whatever_handler_returns() {
+    $fixture= new Invocation(function($req, $res) { return 'It worked!'; }, []);
+    $this->assertEquals('It worked!', $fixture->proceed(new Request(new TestInput('GET', '/')), new Response(new TestOutput())));
+  }
 }